Consolidate web utilities under utils
