+++ packages/web/src/components/documents/DocumentPage.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, useCallback } from 'react'
 import { useParams, Link } from 'react-router-dom'
-import { formatDate } from '../../util/dates.js'
+import { formatDate } from '../../utils/dates.js'
 import { useQuery, useStore } from '@livestore/react'
 import { getDocumentById$ } from '@work-squared/shared/queries'
 import { events } from '@work-squared/shared/schema'
-import { preserveStoreIdInUrl } from '../../util/navigation.js'
+import { preserveStoreIdInUrl } from '../../utils/navigation.js'
 import { MarkdownRenderer } from '../markdown/MarkdownRenderer.js'
 import { LoadingState } from '../ui/LoadingState.js'
 import { LoadingSpinner } from '../ui/LoadingSpinner.js'
