+++ packages/web/src/components/layout/Navigation.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect, useRef } from 'react'
 import { Link, useLocation } from 'react-router-dom'
 import { useQuery } from '@livestore/react'
 import { getUsers$ } from '@work-squared/shared/queries'
-import { getInitials } from '../../util/initials.js'
-import { preserveStoreIdInUrl } from '../../util/navigation.js'
+import { getInitials } from '../../utils/initials.js'
+import { preserveStoreIdInUrl } from '../../utils/navigation.js'
 import { ROUTES, ROUTE_PATTERNS } from '../../constants/routes.js'
 import { useAuth } from '../../contexts/AuthContext.js'
 import { isCurrentUserAdmin } from '../../utils/adminCheck.jsx'
