+++ packages/web/src/components/project-creation/ProjectCreationView.tsx
@@ -14,7 +14,7 @@ import type {
 import { ProjectCreationStage1Presenter } from './ProjectCreationStage1Presenter'
 import { ProjectCreationStage2Presenter } from './ProjectCreationStage2Presenter'
 import { PROJECT_CATEGORIES, type ProjectCategory } from '@work-squared/shared'
-import { preserveStoreIdInUrl } from '../../util/navigation.js'
+import { preserveStoreIdInUrl } from '../../utils/navigation.js'
 import { useCategoryAdvisor } from '../../hooks/useCategoryAdvisor.js'
 
 /**
