+++ packages/web/src/components/projects/ProjectCard/ProjectCard.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { useStore } from '@livestore/react'
-import { formatDate } from '../../../util/dates.js'
+import { formatDate } from '../../../utils/dates.js'
 import type { Project, Worker } from '@work-squared/shared/schema'
 import { getProjectWorkers$, getWorkers$ } from '@work-squared/shared/queries'
 import { getAvatarColor } from '../../../utils/avatarColors.js'
