+++ packages/web/src/components/tasks/TaskModal/TaskModal.tsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react'
 import { useQuery, useStore } from '@livestore/react'
-import { formatDateTime } from '../../../util/dates.js'
+import { formatDateTime } from '../../../utils/dates.js'
 import type { Task, User, Comment } from '@work-squared/shared/schema'
 import { getTaskById$, getUsers$, getTaskComments$ } from '@work-squared/shared/queries'
 import { events } from '@work-squared/shared/schema'
 import { AssigneeSelector } from '../../ui/AssigneeSelector/AssigneeSelector.js'
-import { getInitials } from '../../../util/initials.js'
+import { getInitials } from '../../../utils/initials.js'
 import { useSnackbar } from '../../ui/Snackbar/Snackbar.js'
 import { Markdown } from '../../markdown/Markdown.js'
 import { MoveTaskModal } from '../MoveTaskModal.js'
