+++ packages/web/src/utils/categoryHelpers.ts
