+++ packages/web/src/utils/initials.ts
