+++ packages/web/tests/unit/workerEvents.test.ts
@@ -1,27 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { DEFAULT_MODEL } from '@work-squared/shared'
-import { generateRandomWorkerName } from '../../src/util/workerNames.js'
 import { workerCreated, workerUpdated } from '@work-squared/shared/events'
-
-describe('Worker Name Generation', () => {
-  it('should generate names in the correct format', () => {
-    // Generate multiple names to test format
-    for (let i = 0; i < 10; i++) {
-      const name = generateRandomWorkerName()
-      expect(name).toMatch(/^[A-Z][a-z]+ [A-Z][a-z]+$/)
-    }
-  })
-
-  it('should generate names with reasonable variety', () => {
-    // Generate many names and check we get more than 1 unique name
-    const names = new Set()
-    for (let i = 0; i < 50; i++) {
-      names.add(generateRandomWorkerName())
-    }
-    // With 420 possible combinations, getting at least 10 unique names in 50 tries is very likely
-    expect(names.size).toBeGreaterThan(10)
-  })
-})
 
 describe('Worker Events', () => {
   describe('workerCreated', () => {
